requirements.txt
openpyxl>=3.0.4
google-cloud-monitoring==2.9.1

